added tests as a project on its own in gradle structure
