+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'EcoRun'
 
-include 'desktop', 'core'
+include 'desktop', 'core', 'test'
