--- test/com/hechsmanwilczak/ecorun/test/RegressionTests.java
@@ -1,68 +1,0 @@
-package com.hechsmanwilczak.ecorun.test;
-
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.JUnit4;
-
-
-
-
-@RunWith(JUnit4.class)
-public class RegressionTests {
-
-// in game functions
-@Test
-public void testMovementKeys(){
-    //test UP, LEFT, RIGHT keys for movement
-
-}
-
-@Test
-public void testcollectItems(){
-
-}
-
-@Test
-public void testScoreChanges(){
-    //increase in score
-
-    //decrease in score
-}
-
-@Test
-public void testKillEnemy(){
-
-}
-
-@Test
-public void testLoseLife(){
-
-}
-
-@Test
-public void testFailLevel(){
-    // lost all lives
-
-    // automatic loss
-}
-
-
-// menu functions
-
-@Test
-public void testLevelSelection(){
-
-}
-
-@Test
-public void testRestartGame(){
-
-}
-
-@Test
-public void testPauseButton(){
-
-}
-
-}
